align device and type

+++ act/back_end/verifier.py
@@ -110,7 +110,7 @@ def seed_from_input_specs(spec_layers) -> Bounds:
             center = L.params.get("center")
             eps = L.params.get("eps")
             if center is not None and eps is not None:
-                e = torch.tensor(eps, dtype=center.dtype)
+                e = torch.tensor(eps)
                 return Bounds(center - e, center + e)
     
     # LIN_POLY only -> error
@@ -141,7 +141,7 @@ def add_all_input_specs(globalC: ConSet, input_ids: List[int], spec_layers) -> N
             else:
                 center = L.params["center"]
                 eps = L.params["eps"]
-                e = torch.tensor(eps, dtype=center.dtype)
+                e = torch.tensor(eps)
                 globalC.add_box(-1, input_ids, Bounds(center - e, center + e))
         elif k == InKind.LIN_POLY:
             A, b = L.params["A"], L.params["b"]
